Move transaction update helper class into separate file
